chore: removed PhotoUploader keep ProfilePhotos
